fixes for fragmented NAL units

+++ JSVM/H264Extension/src/lib/H264AVCCommonLib/CFMO.cpp
@@ -302,7 +302,7 @@ void FMO::calcMbNumInSliceGroup()
 	for( i=0; i<NumberOfSliceGroups_; i++)
 		numMbInSliceGroup_[i] = 0;
 
-	for( i=0; i<PicSizeInMapUnits_; i++)
+	for( i=0; i<(Int)PicSizeInMapUnits_; i++)
 		numMbInSliceGroup_[getSliceGroupId(i)]++;
 }
